feat: show albums
