+++ app/javascript/apis/panel.api.ts
@@ -1,0 +1,6 @@
+import type { AxiosResponse } from 'axios';
+import { http } from '@/services/http.service';
+
+export const getAlbums = (): Promise<AxiosResponse> => {
+  return http.get('albums');
+};
