+++ vite.config.ts
@@ -19,4 +19,11 @@ export default defineConfig({
       },
     },
   },
+  server: {
+    hmr: {
+      host: 'localhost',
+      protocol: 'ws',
+      clientPort: 3036,
+    },
+  },
 });
